Refector to move game functionality into its own module.
